Calculate an adjusted SSE/RMSE value.
Add an adjusted error value that tracks the distance from the target's edge
instead of its center.  This seemed to concern students in PSYC 3100 in
SP-2011, so I've added it.  The program still reports the center-based RMSE
for data collection because it is a more sensitive measure.  However,
subjects are now told their adjusted RMSE.