Add windshield-wiper simulation.
Students in PSYC 300 in Spring 2019 wanted to test the distracting effects
of rain with sound effects and something like a windshield wiper.  I added
a black box that briefly blocks the display on a regular cycle.  The code
to support this is included, with a hard-coded enable flag that completely
shuts it off.  The flag is currently off.